Add TodoPorts demo
This demo is the first one where it's starting to feel like it makes
less sense to package all the output as a single HTML file. Previous
demos were just a matter of writing all the source code and then using
`elm-make` to produce an HTML file that worked fine. But in this case
there's a decent amount of code that I ended up manually adding to the
HTML output to get the localStorage feature working. Given that I've
been working towards moving from building minigame demos to more
real-world application ideas, it might be a good idea to split the demos
up into separate categories.